🎨 불필요한 레포지토리 추상 클래스 제거
- 불필요한 xxRepository 추상 클래스를 제거하고 Repository 인터페이스를 바로 구현하도록 구조를 개선한다.
