+++ src/main/java/com/kakao/cafe/repository/VolatilityUserRepository.java
@@ -1,20 +1,49 @@
 package com.kakao.cafe.repository;
 
 import com.kakao.cafe.domain.User;
+import org.springframework.stereotype.Service;
 
+import java.util.Collections;
+import java.util.List;
+import java.util.Optional;
 import java.util.Vector;
 
-public abstract class VolatilityUserRepository implements Repository<User, String> {
+@Service
+public class VolatilityUserRepository implements Repository<User, String> {
 
-    protected final Vector<User> users = new Vector<>();
+    private final Vector<User> users = new Vector<>();
 
-    protected User persist(User user) {
+    @Override
+    public List<User> findAll() {
+        return Collections.unmodifiableList(users);
+    }
+
+    @Override
+    public synchronized Optional<User> save(User user) {
+        Optional<User> other = findOne(user.getUserId());
+
+        User result = other.isEmpty()
+                ? persist(user)
+                : merge(other.get().getIndex(), user);
+
+        return Optional.ofNullable(result);
+    }
+    private User persist(User user) {
         user.setIndex(users.size() + 1);
         return users.add(user) ? user : null;
     }
-
-    protected User merge(int index, User user) {
+    private User merge(int index, User user) {
         users.set(index - 1, user);
         return user;
+    }
+
+    @Override
+    public Optional<User> findOne(String id) {
+        return users.stream().filter(user -> user.ownerOf(id)).findAny();
+    }
+
+    @Override
+    public void clear() {
+        users.clear();
     }
 }
