+++ src/test/java/com/kakao/cafe/service/VolatilityArticleServiceTest.java
@@ -17,7 +17,7 @@ import static org.mockito.BDDMockito.given;
 import static org.mockito.Mockito.verify;
 
 @ExtendWith(MockitoExtension.class)
-public class VolatilityArticleServiceUnitTest {
+public class VolatilityArticleServiceTest {
 
     @InjectMocks
     VolatilityArticleService articleService;
